Universal-Arduino-Display-Simulator/requirements.txt
pygame>=2.5.0
pyserial>=3.5